test(logic): split connectivity blockade tests under 400 lines
Refs #2288

Remove the monolithic connectivity_resolver_blockade_test.dart and add
three segment files under test/world/, preserving the same groups and
assertions. Relocates coverage next to other world connectivity tests.
